Move metal tests back to gpu.
