+++ gpu/Cargo.toml
@@ -27,6 +27,7 @@ default = []
 [dev-dependencies]
 proptest = "1.1.0"
 criterion = "0.4.0"
+lambdaworks-fft = { path = "../fft" }
 
 [[bench]]
 name = "metal_benchmarks"
